+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Connection to local database
-- Command line parameters to configure connection to local database (-d for database enabled, -r for database port, -u for database username, -w for database password)
+- Connection to local database, supporting Postgresql and sqlite
+- Command line parameters to configure connection to local database (-d for database enabled, -e for database engine, -r for database port, -u for database username, -w for database password)
 - Load command
 
 ### Changed